requirements.txt
requests~=2.26.0
matplotlib~=3.4.3
Pillow~=8.1.1
PyQt5~=5.15.4
bs4~=0.0.1
beautifulsoup4~=4.10.0